art_gallery/requirements.txt
Django==5.1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Pillow==10.2.0
mysqlclient==2.2.4
python-dotenv==1.0.1 